pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "u1kit"
version = "0.1.0"
description = "Convert Bambu/Makerworld .3mf files for the Snapmaker U1"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "prompt-toolkit>=3.0",
    "platformdirs>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "mypy>=1.8,<1.15",
    "ruff>=0.5",
    "types-PyYAML>=6.0",
]
# Only needed to build the PyInstaller sidecar for the Phase 3 GUI; not
# required for regular dev work. Install with `pip install -e ".[release]"`.
release = [
    "pyinstaller>=6.5",
]

[project.scripts]
u1kit = "u1kit.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["u1kit"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
strict = true
python_version = "3.10"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM"]
